requirements.txt
fastapi==0.116.1
httpx==0.28.1
matplotlib==3.10.6
numpy==2.3.3
Pillow==11.3.0
playwright==1.55.0
protobuf==6.32.1
pydantic==2.11.9
pyngrok==7.3.0
python-dotenv==1.1.1
tqdm==4.67.1
uvicorn==0.35.0
torchvision==0.23.0
webdataset